Allow user to configure SAAINT to use a fixed seed (#7)
for the random number generator when an adventure is started or
restored.

This is specified using the "seed" key in the .ini file. The
value must be 0 or a positive integer (anything else is replaced
by 0). The default behaviour (seed = 0) is for the random number
generator to be seeded from the TIMER as in the previous version.

Closes-Bug: #7