Stop rooms being redescribed unnecessarily
This was visible in 'Ghost Town' where the original location was being
described twice because an automatic action was (for some reason) moving
the player from and then back to the current location thus causing
redraw_flag% to be set.

redraw_flag% may now be superflous as anything other than an
optimisation. The new interp.has_changed%() function can robustly
determine if a room needs (re)describing without reference to this flag.