Yet another attempt at sorting out room redescribing logic
Tested and working for:
  SACA 1: Adventureland
  SACA 2: Pirate Adventure (BYTE Magazine)
  SACA 3: Secret Mission
  SACA 8: Pyramid of Doom
  The Great Blackberry Adventure

The biggest sticking point was the behaviour of the "small nomad"
in "Pyramid of Doom". This is because if a room is automatically
described when the player moves into it and also when its content
changes then the process of the nomad following the player around
causes rooms to be described twice, one with, and once without the
nomad.

Closes-Bug: #9, #10